+++ app/src/main/java/com/globant/openbankassignment/domain/repository/GetCharactersDetailsRepository.kt
@@ -4,5 +4,5 @@ import com.globant.openbankassignment.data.entity.MarvelCharactersResponse
 import io.reactivex.Observable
 
 interface GetCharactersDetailsRepository {
-    fun getCharactersDetailsById(characterId:Int): Observable<MarvelCharactersResponse>
+    fun getCharactersDetailsById(characterId:Long): Observable<MarvelCharactersResponse>
 }
